+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- [TD-6535] Delete elasticsearch index only with the index name
+
 ## [6.5.1] 2024-04-15
 
 ### Added